Add a code visualisation for regular expressions
The visualisation is based on Regulex, an online tool for visualising regexes. The code is open-source, but seems little portable as is, so the visualisation is provided through <iframe> elements instead. It is a little buggy and requires some strange heuristics (e.g., to size/scale the element), and there is a bug regarding updates on document changes, but it already gives a rough idea of what such a visualisation could provide.

The TypeScript example has been updated to add two regexes (one is a literal, and the other one uses the RegExp constructor).