Improve monocle renderers
The renderers have been refactored and improved (especially in terms of positioning).

In addition, in order to keep a monocle active (and therefore displayed by its renderer) even though monocles are re-created after each non-transient edit, they now have a transferable state.

It is not set when and where it should yet, but this is the first step to transfer information accross different generations of monocles!