Add support for runtime requests/responses
Runtime requests are a way for monocles to query information available at runtime (such as the value of a variable) that can be then be used within the monocle. In the VS Code editor app, requests and responses are transferred to/received from the extension, which is responsible for actually getting the runtime information (e.g., from the debugger).