Add utilities to specify and test values
A Valuable can be used to represent a value that is either passed directly or returned by evaluating a function.

A ValueCondition can be used to represent a condition on a value that either compares it to another value, tests if a list of values contains it, or evaluate an arbitrary predicate function with it.